add assemble.c
